+++ Core/changelog.md
@@ -1,5 +1,24 @@
 # MyTrout.Pipelines.Core Change Log
 
+
+## 4.0.0 = BREAKING CHANGES
+- Alter IStepActivator interface to include the ParameterCreators property to allow developers to reconfigure the parameter creation behavior.
+- BREAKING CHANGE: Remove support for .NET 5.0.
+- BREAKING CHANGE: Move ParameterCreationDelegate and ParameterCreationResult to the MyTrout.Pipelines namespace due to IStepActivator.ParameterCreators change to preserve namespace dependencies.
+- BREAKING CHANGE: Alter ParameterCreationDelegate to use ILogger<IStepActivator> instead of ILogger<StepActivator> to preserve namespace dependencies.
+- BREAKING CHANGE: Alter StepActivator delegate methods to use ILogger<IStepActivator> instead of ILogger<StepActivator> to preserve namespace dependencies.
+- Alter ParameterCreators property use ParameterCreationDelegate that injects both IConfiguration and Dependency Injection values.
+- Uncomment all of the nuget publish steps to allow a new version to be published.
+- Mark the Configuration and IsConfigurationAvailable properties with the Obsolete attribute as the new LoadConfigurationToPipelineContextStep supercedes this functionality.
+- Add CreateUnixEpochStep.
+- Add LoadValuesFromContextObjectToPipelineContextStep<TObject>
+- Add LoadValuesFromCConfigurationToPipelineContextStep.
+- Add UnixEpochKind to support different epoch configurations in CreateUnixEpochStep.
+- Add INPUT_OBJECT, OUTPUT_OBJECT, and UNIX_EPOCH to PipelineContextConstants.
+- Change MoveOutputStreamToInputStreamStep and MoveInputStreamToOutputStream to use the RenameContextItemStep as the base class, thereby eliminating code.
+- Correct documentation typos on RenameContextItemStep.
+- Add support for .NET 7.0
+
 ## 3.2.0 - SonarCloud UPDATE ONLY
 - Suppress CA2254 to prevent false positives in SonarCloud.io with culture-aware logging messages
 - Correct test project steps to conform with CA1816 warning